Added unit tests for document processing service

+++ src/main/java/com/document/document/message/DocumentMessage.java
@@ -6,19 +6,19 @@ import java.io.Serializable;
 
 public class DocumentMessage implements Serializable {
     @JsonProperty("id")
-    Integer id;
+    Long id;
 
     @JsonProperty("content")
     String content;
 
     DocumentMessage() {}
 
-    public DocumentMessage(Integer id, String content) {
+    public DocumentMessage(Long id, String content) {
         this.id = id;
         this.content = content;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
